requirements.txt
pre-commit==4.2.0
pytest==8.4.0
requests==2.32.4
python-dotenv==1.1.0
jsonschema==4.24.0}
Faker==37.4.0
pytest-html==4.1.1
pytest-excel==1.7.0
openpyxl==3.1.5
pytest-md-report==0.7.0
allure-pytest==2.14.3
pymsteams==0.2.5
